New Module
Add new module, a web server to view the log files more conveniently.
This module allows you to see the four main files of this bot through a web page. Once the bot and this module have started, we can access the page http: // localhost: 8000 / from our favorite browser and see choose the live mode or the simulation mode and the history, bot_stats, trades and coins_bought files will be displayed.
Clarification, if any of these files do not exist or have not yet been created, it may show you an error in its presentation socket.